fix:follow logic moved to "FollowButton"
